pyproject.toml
[tool.poetry]
name = "dashed"
version = "0.1.0"
description = ""
authors = ["andrei <b1naryth1ef@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
